ColumnImpl moved into its own header file
