+++ c/column.cc
@@ -12,6 +12,7 @@
 #include "utils/file.h"
 #include "utils/misc.h"
 #include "column.h"
+#include "column_impl.h"
 #include "datatablemodule.h"
 #include "rowindex.h"
 #include "sort.h"
