+++ c/column_fw.cc
@@ -10,6 +10,7 @@
 #include "utils/misc.h"
 #include "parallel/api.h"  // dt::parallel_for_static
 #include "column.h"
+#include "column_impl.h"
