+++ c/expr/expr_reduce.cc
@@ -11,6 +11,7 @@
 #include <unordered_map>      // std::unordered_map
 #include "expr/expr_reduce.h"
 #include "parallel/api.h"
+#include "column_impl.h"  // TODO: remove
 #include "types.h"
 namespace dt {
 namespace expr {
