+++ c/expr/expr_unaryop.cc
@@ -9,6 +9,7 @@
 #include "expr/expr_unaryop.h"
 #include "frame/py_frame.h"
 #include "parallel/api.h"
+#include "column_impl.h"  // TODO: remove
 #include "datatablemodule.h"
 #include "types.h"
 namespace dt {
