+++ c/expr/repl_node.cc
@@ -25,6 +25,7 @@
 #include "expr/repl_node.h"
 #include "expr/workframe.h"
 #include "utils/exceptions.h"
+#include "column_impl.h"  // TODO: remove
 #include "datatable.h"
 #include "datatablemodule.h"
 namespace dt {
