+++ c/frame/__init__.cc
@@ -28,6 +28,7 @@
 #include "python/oset.h"
 #include "python/string.h"
 #include "utils/alloc.h"
+#include "column_impl.h"  // TODO: remove
 #include "ztest.h"
 
 namespace py {
