+++ c/frame/cbind.cc
@@ -19,6 +19,7 @@
 #include "utils/assert.h"
 #include "datatable.h"
 #include "datatablemodule.h"
+#include "column_impl.h"  // TODO: remove
 
 namespace py {
