+++ c/models/label_encode.h
@@ -23,6 +23,7 @@
 #include "parallel/shared_mutex.h"
 #include "utils/exceptions.h"
 #include "models/dt_ftrl_base.h"
+#include "column_impl.h"  // TODO: remove
 namespace dt {
