+++ c/str/split_into_nhot.cc
@@ -23,6 +23,7 @@
 #include "str/py_str.h"
 #include "datatable.h"
 #include "options.h"
+#include "column_impl.h"  // TODO: remove
 
 namespace dt {
